requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
boto3>=1.34.0
opensearch-py>=2.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
